requirements.txt
google-adk>=1.29.0
google-cloud-bigquery
google-cloud-storage
google-cloud-documentai
google-cloud-discoveryengine
fastapi
uvicorn
requests
